Rename misleadingly named primitive builder implementations.
