+++ core/src/main/java/io/atomix/core/map/impl/DefaultAtomicMapBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Open Networking Foundation
+ * Copyright 2016-present Open Networking Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,14 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package io.atomix.core.multimap.impl;
+package io.atomix.core.map.impl;
 
 import com.google.common.io.BaseEncoding;
-import io.atomix.core.multimap.AsyncAtomicMultimap;
-import io.atomix.core.multimap.AtomicMultimap;
-import io.atomix.core.multimap.AtomicMultimapBuilder;
-import io.atomix.core.multimap.AtomicMultimapConfig;
+import io.atomix.core.map.AsyncAtomicMap;
+import io.atomix.core.map.AtomicMap;
+import io.atomix.core.map.AtomicMapBuilder;
+import io.atomix.core.map.AtomicMapConfig;
 import io.atomix.primitive.PrimitiveManagementService;
 import io.atomix.primitive.proxy.ProxyClient;
 import io.atomix.primitive.service.ServiceConfig;
@@ -29,36 +28,48 @@ import io.atomix.utils.serializer.Serializer;
 import java.util.concurrent.CompletableFuture;
 
 /**
- * Default {@link AsyncAtomicMultimap} builder.
+ * Default {@link AsyncAtomicMap} builder.
+ *
+ * @param <K> type for map key
+ * @param <V> type for map value
  */
-public class AtomicMultimapProxyBuilder<K, V> extends AtomicMultimapBuilder<K, V> {
-  public AtomicMultimapProxyBuilder(String name, AtomicMultimapConfig config, PrimitiveManagementService managementService) {
+public class DefaultAtomicMapBuilder<K, V> extends AtomicMapBuilder<K, V> {
+  public DefaultAtomicMapBuilder(String name, AtomicMapConfig config, PrimitiveManagementService managementService) {
     super(name, config, managementService);
   }
 
   @Override
   @SuppressWarnings("unchecked")
-  public CompletableFuture<AtomicMultimap<K, V>> buildAsync() {
-    ProxyClient<AtomicMultimapService> proxy = protocol().newProxy(
+  public CompletableFuture<AtomicMap<K, V>> buildAsync() {
+    ProxyClient<AtomicMapService> proxy = protocol().newProxy(
         name(),
         primitiveType(),
-        AtomicMultimapService.class,
+        AtomicMapService.class,
         new ServiceConfig(),
         managementService.getPartitionService());
-    return new AtomicMultimapProxy(proxy, managementService.getPrimitiveRegistry())
+    return new AtomicMapProxy(proxy, managementService.getPrimitiveRegistry())
         .connect()
-        .thenApply(rawMultimap -> {
+        .thenApply(rawMap -> {
           Serializer serializer = serializer();
-          AsyncAtomicMultimap<K, V> multimap = new TranscodingAsyncAtomicMultimap<>(
-              rawMultimap,
+          AsyncAtomicMap<K, V> map = new TranscodingAsyncAtomicMap<K, V, String, byte[]>(
+              rawMap,
               key -> BaseEncoding.base16().encode(serializer.encode(key)),
               string -> serializer.decode(BaseEncoding.base16().decode(string)),
               value -> serializer.encode(value),
               bytes -> serializer.decode(bytes));
-          if (config.isCacheEnabled()) {
-            multimap = new CachingAsyncAtomicMultimap<>(multimap);
+
+          if (!config.isNullValues()) {
+            map = new NotNullAsyncAtomicMap<>(map);
           }
-          return multimap.sync();
+
+          if (config.isCacheEnabled()) {
+            map = new CachingAsyncAtomicMap<>(map, config.getCacheSize());
+          }
+
+          if (config.isReadOnly()) {
+            map = new UnmodifiableAsyncAtomicMap<>(map);
+          }
+          return map.sync();
         });
   }
 }
