+++ core/src/main/java/io/atomix/core/set/DistributedSetType.java
@@ -19,7 +19,7 @@ import io.atomix.core.collection.CollectionEvent;
 import io.atomix.core.collection.impl.CollectionUpdateResult;
 import io.atomix.core.set.impl.DefaultDistributedSetService;
 import io.atomix.core.collection.impl.DistributedCollectionService;
-import io.atomix.core.set.impl.DistributedSetProxyBuilder;
+import io.atomix.core.set.impl.DefaultDistributedSetBuilder;
 import io.atomix.core.set.impl.DistributedSetResource;
 import io.atomix.core.set.impl.SetUpdate;
 import io.atomix.core.transaction.TransactionId;
@@ -99,7 +99,7 @@ public class DistributedSetType<E> implements PrimitiveType<DistributedSetBuilde
 
   @Override
   public DistributedSetBuilder<E> newBuilder(String name, DistributedSetConfig config, PrimitiveManagementService managementService) {
-    return new DistributedSetProxyBuilder<>(name, config, managementService);
+    return new DefaultDistributedSetBuilder<>(name, config, managementService);
   }
 
   @Override
