+++ core/src/main/java/io/atomix/core/workqueue/impl/DefaultWorkQueueBuilder.java
@@ -13,37 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.barrier.impl;
+package io.atomix.core.workqueue.impl;
 
-import io.atomix.core.barrier.AsyncDistributedCyclicBarrier;
-import io.atomix.core.barrier.DistributedCyclicBarrier;
-import io.atomix.core.barrier.DistributedCyclicBarrierBuilder;
-import io.atomix.core.barrier.DistributedCyclicBarrierConfig;
+import io.atomix.core.workqueue.WorkQueue;
+import io.atomix.core.workqueue.WorkQueueBuilder;
+import io.atomix.core.workqueue.WorkQueueConfig;
 import io.atomix.primitive.PrimitiveManagementService;
 import io.atomix.primitive.proxy.ProxyClient;
 import io.atomix.primitive.service.ServiceConfig;
+import io.atomix.utils.serializer.Serializer;
 
 import java.util.concurrent.CompletableFuture;
 
 /**
- * Default distributed cyclic barrier builder implementation.
+ * Default work queue builder implementation.
  */
-public class DistributedCyclicBarrierProxyBuilder extends DistributedCyclicBarrierBuilder {
-  public DistributedCyclicBarrierProxyBuilder(String name, DistributedCyclicBarrierConfig config, PrimitiveManagementService managementService) {
+public class DefaultWorkQueueBuilder<E> extends WorkQueueBuilder<E> {
+  public DefaultWorkQueueBuilder(String name, WorkQueueConfig config, PrimitiveManagementService managementService) {
     super(name, config, managementService);
   }
 
   @Override
   @SuppressWarnings("unchecked")
-  public CompletableFuture<DistributedCyclicBarrier> buildAsync() {
-    ProxyClient<DistributedCyclicBarrierService> proxy = protocol().newProxy(
+  public CompletableFuture<WorkQueue<E>> buildAsync() {
+    ProxyClient<WorkQueueService> proxy = protocol().newProxy(
         name(),
         primitiveType(),
-        DistributedCyclicBarrierService.class,
+        WorkQueueService.class,
         new ServiceConfig(),
         managementService.getPartitionService());
-    return new DistributedCyclicBarrierProxy(proxy, managementService.getPrimitiveRegistry(), barrierAction)
+    return new WorkQueueProxy(proxy, managementService.getPrimitiveRegistry())
         .connect()
-        .thenApply(AsyncDistributedCyclicBarrier::sync);
+        .thenApply(queue -> {
+          Serializer serializer = serializer();
+          return new TranscodingAsyncWorkQueue<E, byte[]>(
+              queue,
+              item -> serializer.encode(item),
+              bytes -> serializer.decode(bytes))
+              .sync();
+        });
   }
 }
